+++ src/Xreferee/Lsp/AppM.hs
@@ -75,7 +75,6 @@ data AppState = AppState
     isDbDirty :: Bool,
     -- | Keep track of which files have warnings/errors.
     filesWithDiagnostics :: Set Uri,
-    fileVersions :: SM.Map Uri Int32,
     -- | Keep track of which files are ignored, see @(ref:shouldHandleFileOrDir)
     shouldHandleFiles :: SM.Map Uri Bool
   }