+++ src/Xreferee/Lsp/Handlers/DidChangeWatchedFiles.hs
@@ -93,10 +93,7 @@ runHandler fileEvents = do
                   pure ()
                 Right contents -> do
                   Log.debug $ "didChangeWatchedFiles: Changed: reloading file from disk: " <> uri.getUri
-                  -- NOTE: If a file is changed on disk (e.g. with `echo "#\(ref:test4)" >> file.md`), AND the file is not currently opened in vscode,
-                  -- the next time the user opens it, the version will be reset to 1.
-                  let fileVersion = 1
-                  Symbols.loadSymbolsForFile uri contents fileVersion
+                  Symbols.loadSymbolsForFile uri contents
           LSP.FileChangeType_Created -> do
             -- NOTE: this is triggered when:
             --  * a file is created via the editor (we receive a `didOpen` notification followed by a `didChangeWatchedFiles`).
@@ -137,8 +134,7 @@ runHandler fileEvents = do
                     pure ()
                   Right contents -> do
                     Log.debug $ "didChangeWatchedFiles: Created: loading file from disk: " <> tshow path
-                    let fileVersion = 1
-                    Symbols.loadSymbolsForFile uri contents fileVersion
+                    Symbols.loadSymbolsForFile uri contents
           LSP.FileChangeType_Deleted -> do
             -- NOTE: We don't know whether this was a file or a directory.
             -- So we have to delete the symbols for this uri, and also delete the symbols for all files with